+++ ide/coqide_ui.ml
@@ -140,6 +140,8 @@ let init () =
 \n    <menuitem action='Uncomment' />\
 \n    <separator />\
 \n    <menuitem action='Coqtop arguments' />\
+\n    <separator />\
+\n    <menuitem action='LaTeX-to-unicode' />\
 \n  </menu>\
 \n  <menu action='Compile'>\
 \n    <menuitem action='Compile buffer' />\